Implement side nav for response-list
